Array initialization now works with the abstract object
Made the constructor for making an abstract object from an expression
also take an abstract_enviromentt and a namespacet so can call
environment.eval on parts of the expression (e.g. to get an abstract
object for each entry in the array).

Use at rather than the operator[] since const correct.