Explained why top must be used as the else clause in eval
I attempted to change this to generating a abstract object directly from
the simplified expression. This caused the tests to fail in what
appeared to be a "better" way - that is the constant_propagator tests
were reporting more assignment simplifications. Specifically, it was
simplifying the return of the function to be a NONDET(int) rather than
the return of the function.

I've convinced the change is wrong and works by chance, if we could some
how have exprt with no children but wasn't a symbol, but also wasn't
really a constant, then this would be wrong. The optimisation is also
probably wrong since if the return assignment happened earlier, the
return value should be the same, not a new non-det.