Enable the write stack for pointers to structs
This works correctly so we don't need to cautiously throw away stacks of
pointers to structs. The original concern was pointers to structs being
used as pointers to the first component, but the goto program introduces
a cast which means we don't create the pointer.

Added a unit test to document this behavior.