Expand full_struct_abstract_objectt construction to initialise struct members
Previously, when constructing a full_struct_abstract_objectt from an
exprt representing a struct value, no information was recorded about
the values of the struct, and the 'last written location' information
was only recorded at the top level of the struct, and not at the level
of it's member elements. This commit ensures that member elements also
get initialised.